simplify kube config watch
